Fix cache loc
