+++ docs/api/model.rst
@@ -20,7 +20,6 @@ Language Modeling
     standard_lstm_lm_1500
     StandardRNN
     get_model
-    get_cache_model
 
 Attention Cell
 --------------
